requirements.txt
asgiref==3.2.10
backcall==0.2.0
colorama==0.4.4
coverage==5.3
decorator==4.4.2
Django==3.1.1
gunicorn==20.0.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
psycopg2-binary==2.9.3
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
traitlets==5.0.5
wcwidth==0.2.5
